+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- **Interactive Menu Enhancements**: Version number display in interactive menu banner
+
 ## [0.9.1] - 2025-11-16
 
 ### Added